+++ R/CreateSpacetimeFolds.R
@@ -35,7 +35,7 @@
 #' str(indices)
 #' ### Prepare for leave-One-Location-Out cross validation
 #' indices <- CreateSpacetimeFolds(dat,spacevar="SOURCEID",
-#' k=length(unique(dat$SOURCEID)))
+#'     k=length(unique(dat$SOURCEID)))
 #' str(indices)
 #' @export CreateSpacetimeFolds
 #' @aliases CreateSpacetimeFolds