+++ R/calibrate_aoa.R
@@ -11,9 +11,9 @@
 #' @param m Numeric. See mgcv::s
 #' @param showPlot Logical.
 #' @details If multiCV=TRUE the model is re-fitted and validated by length.out new cross-validations where the cross-validation folds are defined by clusters in the predictor space,
-#' ranging from three clusters to LOOCV.
-#' If the AOA threshold based on the calibration data from multiple CV is larger than the original AOA threshold, the AOA is updated accordingly.
-#' See Meyer and Pebesma (2020) for the full documentation of the methodology.
+#' ranging from three clusters to LOOCV. Hence, a large range of DI values is created during cross-validation.
+#' If the AOA threshold based on the calibration data from multiple CV is larger than the original AOA threshold (which is likely if extrapolation situations are created during CV),
+#' the AOA is updated accordingly. See Meyer and Pebesma (2021) for the full documentation of the methodology.
 #' @return A list of length 2 with the elements "AOA": rasterStack which contains the original DI and the AOA (which might be updated if new test data indicate this option), as well as the expected performance based on the relationship.
 #' Data used for calibration are stored in the attributes. The second element is a plot showing the relationship.
 #' @author
@@ -55,7 +55,8 @@
 #'
 # '# and get the expected performance on a pixel-level:
 #' AOA_new <- calibrate_aoa(AOA,model)
-#' plot(AOA_new$AOA[[3]])
+#' plot(AOA_new$AOA$expected_RMSE)
+#' # attributes(AOA_new$AOA) # data used for calibration
 #' }
 #' @export calibrate_aoa
 #' @aliases calibrate_aoa